WIP spliting crates
