+++ kindelia_core/src/NoHashHasher.rs
