+++ kindelia_core/src/common.rs
