+++ kindelia_core/src/constants.rs
@@ -2,4 +2,4 @@
 // ============
 
 /// Kdl code included on genesis block
-pub const GENESIS_CODE: &str = include_str!("../genesis.kdl");
+pub const GENESIS_CODE: &str = include_str!("genesis.kdl");
