+++ kindelia_core/src/hvm.rs
