+++ kindelia_core/src/persistence.rs
