+++ kindelia_core/src/test/bits.rs
