+++ kindelia_core/src/test/hasher.rs
