+++ kindelia_core/src/test/hvm.rs
