+++ kindelia_core/src/test/network.rs
