+++ kindelia_core/src/test/node.rs
