+++ kindelia_core/src/test/strategies.rs
