+++ kindelia_core/src/test/util.rs
