+++ kindelia_core/src/util.rs
